requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0